Use explicit output name to avoid the output file being called m
The deployment script expects the game name to match the repo/folder
name but the go mod file calls it m, so the resulting binary is m and
the archive program can't find it because it's looking for "attraction".

Specifying -o attraction to go build solves this and now it's
consistently called attraction everywhere.